effekseer移植実装、とりあえず一通り動くようになった 実装上、パイプラインの生成に失敗する場所があるので原因を特定する next:テクスチャリソースの読込 RootSignatureをRenderStateからShaderResourceに移植

+++ EffekseerOn12/EffekseerRuntime130/src/EffekseerRendererDX12/EffekseerRendererDX12.IndexBuffer.cpp
@@ -55,7 +55,7 @@ namespace EffekseerRendererDX12
 		D3D12_INDEX_BUFFER_VIEW ibv;
 		ibv.BufferLocation = ib->GetGPUVirtualAddress();
 		ibv.Format = DXGI_FORMAT_R16_UINT;
-		ibv.SizeInBytes = hBufferDesc.Width;
+		ibv.SizeInBytes = (UINT)hBufferDesc.Width;
 
 		return new IndexBuffer(renderer, ib, ibv, maxCount, isDynamic);
 	}
